Extract encryption method to a class
