+++ lib/fake_idp/application.rb
@@ -1,11 +1,8 @@
-require "xmlenc"
+require_relative "encryptor"
 
 module FakeIdp
   class Application < Sinatra::Base
     include SamlIdp::Controller
-
-    ENCRYPTION_STRATEGY = "aes256-cbc".freeze
-    KEY_TRANSPORT = "rsa-oaep-mgf1p".freeze
 
     get '/saml/auth' do
       begin
@@ -108,79 +105,14 @@ module FakeIdp
       signature = %[<ds:Signature xmlns:ds="http://www.w3.org/2000/09/xmldsig#">#{signed_info}<ds:SignatureValue>#{signature_value}</ds:SignatureValue><KeyInfo xmlns="http://www.w3.org/2000/09/xmldsig#"><ds:X509Data><ds:X509Certificate>#{self.x509_certificate}</ds:X509Certificate></ds:X509Data></KeyInfo></ds:Signature>]
 
       assertion_and_signature = assertion.sub(/Issuer\>\<saml:Subject/, "Issuer>#{signature}<saml:Subject")
-      assertion_and_signature = encrypt(assertion_and_signature) if configuration.encryption_enabled
+
+      if configuration.encryption_enabled
+        assertion_and_signature = Encryptor.new(assertion_and_signature, idp_certificate).encrypt
+      end
 
       xml = %[<samlp:Response ID="_#{response_id}" Version="2.0" IssueInstant="#{now.iso8601}" Destination="#{@saml_acs_url}" Consent="urn:oasis:names:tc:SAML:2.0:consent:unspecified"#{@saml_request_id ? %[ InResponseTo="#{@saml_request_id}"] : ""} xmlns:samlp="urn:oasis:names:tc:SAML:2.0:protocol"><saml:Issuer xmlns:saml="urn:oasis:names:tc:SAML:2.0:assertion">#{issuer_uri}</saml:Issuer><samlp:Status><samlp:StatusCode Value="urn:oasis:names:tc:SAML:2.0:status:Success" /></samlp:Status>#{assertion_and_signature}</samlp:Response>]
 
       Base64.encode64(xml)
-    end
-
-    # Encryption approach borrowed from
-    # https://github.com/saml-idp/saml_idp/blob/master/lib/saml_idp/encryptor.rb
-    def encrypt(raw_xml) 
-      encryption_template = Nokogiri::XML::Document.parse(build_encryption_template).root
-      encrypted_data = Xmlenc::EncryptedData.new(encryption_template)
-      encryption_key = encrypted_data.encrypt(raw_xml)
-      encrypted_key_node = encrypted_data.node.at_xpath(
-        "//xenc:EncryptedData/ds:KeyInfo/xenc:EncryptedKey",
-        Xmlenc::NAMESPACES
-      )   
-      encrypted_key = Xmlenc::EncryptedKey.new(encrypted_key_node)
-      encrypted_key.encrypt(openssl_cert.public_key, encryption_key)
-
-      xml = Builder::XmlMarkup.new
-      xml.EncryptedAssertion xmlns: "urn:oasis:names:tc:SAML:2.0:assertion" do |enc_assert|
-        enc_assert << encrypted_data.node.to_s
-      end 
-    end
-
-    def openssl_cert
-      if idp_certificate.is_a?(String)
-        @_openssl_cert ||= OpenSSL::X509::Certificate.new(Base64.decode64(idp_certificate))
-      else
-        @_openssl_cert ||= idp_certificate
-      end 
-    end
-
-    def encryption_strategy_ns
-      "http://www.w3.org/2001/04/xmlenc##{ENCRYPTION_STRATEGY}"
-    end
-
-    def key_transport_ns
-      "http://www.w3.org/2001/04/xmlenc##{KEY_TRANSPORT}"
-    end
-
-    def build_encryption_template
-      xml = Builder::XmlMarkup.new
-      xml.EncryptedData Id: "ED", Type: "http://www.w3.org/2001/04/xmlenc#Element",
-        xmlns: "http://www.w3.org/2001/04/xmlenc#" do |enc_data|
-        enc_data.EncryptionMethod Algorithm: encryption_strategy_ns
-        enc_data.tag! "ds:KeyInfo", "xmlns:ds" => "http://www.w3.org/2000/09/xmldsig#" do |key_info|
-          key_info.EncryptedKey Id: "EK", xmlns: "http://www.w3.org/2001/04/xmlenc#" do |enc_key|
-            enc_key.EncryptionMethod Algorithm: key_transport_ns
-            enc_key.tag! "ds:KeyInfo", "xmlns:ds" => "http://www.w3.org/2000/09/xmldsig#" do |key_info_child|
-              key_info_child.tag! "ds:KeyName"
-              key_info_child.tag! "ds:X509Data" do |x509_data|
-                x509_data.tag! "ds:X509Certificate" do |x509_cert|
-                  x509_cert << idp_certificate.to_s.gsub(/-+(BEGIN|END) CERTIFICATE-+/, "") 
-                end
-              end
-            end
-
-            enc_key.CipherData do |cipher_data|
-              cipher_data.CipherValue
-            end
-
-            enc_key.ReferenceList do |ref_list|
-              ref_list.DataReference URI: "#ED"
-            end
-          end
-        end
-
-        enc_data.CipherData do |cipher_data|
-          cipher_data.CipherValue
-        end
-      end
     end
   end
 end
